+++ .eslintrc.js
@@ -12,11 +12,12 @@ module.exports = {
     node: true,
     jest: true,
   },
-  ignorePatterns: ['.eslintrc.js', 'rollup.config.js'],
+  ignorePatterns: ['.eslintrc.js', 'rollup.config.js', 'examples/*'],
   rules: {
     '@typescript-eslint/interface-name-prefix': 'off',
     '@typescript-eslint/explicit-function-return-type': 'off',
     '@typescript-eslint/explicit-module-boundary-types': 'off',
     '@typescript-eslint/no-explicit-any': 'off',
+    'no-console': ['error', { allow: ['warn', 'error'] }],
   },
 };